requirements.txt
asgiref==3.2.3
Django==2.0.5
djangorestframework==3.11.0
#python 3.8
# mysqlclient==2.0.3
#python 3.6
mysqlclient==1.3.12
pytz==2019.3
sqlparse==0.3.0
django-oauth-toolkit==1.1.1
html2text==2020.1.16
django-jsonfield==1.4.0
paypalrestsdk==1.13.1
boto3==1.14.63
timeago==1.0.14
django-environ
django-rest-swagger==2.2.0